Updating to webpack 4
