+++ package.json
@@ -1,106 +1,54 @@
 {
-  "name": "react-box",
-  "version": "0.1.0",
-  "private": true,
+  "name": "gnosis-team-safe",
+  "version": "1.0.0",
+  "description": "Allowing crypto users manage funds in a safer way",
+  "directories": {
+    "test": "test"
+  },
+  "scripts": {
+    "start": "webpack --mode development --config config/webpack.dev.js",
+    "build": "webpack --mode production --config config/webpack.dev.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/gnosis/gnosis-team-safe"
+  },
+  "author": "Gnosis Team",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/gnosis/gnosis-team-safe/issues"
+  },
+  "homepage": "https://github.com/gnosis/gnosis-team-safe#readme",
   "devDependencies": {
-    "autoprefixer": "6.5.1",
-    "babel-core": "6.17.0",
-    "babel-eslint": "7.1.1",
-    "babel-jest": "17.0.2",
-    "babel-loader": "6.2.7",
-    "babel-preset-react-app": "^2.0.1",
-    "case-sensitive-paths-webpack-plugin": "1.1.4",
-    "chalk": "1.1.3",
-    "connect-history-api-fallback": "1.3.0",
-    "cross-spawn": "4.0.2",
-    "css-loader": "0.26.0",
-    "detect-port": "1.0.1",
-    "dotenv": "2.0.0",
-    "eslint": "3.8.1",
-    "eslint-config-react-app": "^0.5.0",
-    "eslint-loader": "1.6.0",
-    "eslint-plugin-flowtype": "2.21.0",
-    "eslint-plugin-import": "2.0.1",
-    "eslint-plugin-jsx-a11y": "2.2.3",
-    "eslint-plugin-react": "6.4.1",
-    "extract-text-webpack-plugin": "1.0.1",
-    "file-loader": "0.9.0",
-    "filesize": "3.3.0",
-    "fs-extra": "0.30.0",
-    "gzip-size": "3.0.0",
-    "html-webpack-plugin": "2.24.0",
-    "http-proxy-middleware": "0.17.2",
-    "jest": "18.1.0",
-    "json-loader": "0.5.4",
-    "object-assign": "4.1.0",
-    "path-exists": "2.1.0",
-    "postcss-loader": "1.0.0",
-    "postcss-simple-vars": "^4.0.0",
-    "promise": "7.1.1",
-    "react-dev-utils": "^0.4.2",
-    "recursive-readdir": "2.1.0",
-    "strip-ansi": "3.0.1",
-    "style-loader": "0.13.1",
-    "truffle-contract": "^1.1.8",
+    "@babel/cli": "^7.0.0-beta.40",
+    "@babel/core": "^7.0.0-beta.40",
+    "@babel/preset-env": "^7.0.0-beta.40",
+    "@babel/preset-react": "^7.0.0-beta.40",
+    "@babel/preset-stage-0": "^7.0.0-beta.40",
+    "autoprefixer": "^8.1.0",
+    "babel-core": "^7.0.0-bridge.0",
+    "babel-loader": "^8.0.0-beta.0",
+    "babel-plugin-transform-es3-member-expression-literals": "^6.22.0",
+    "babel-plugin-transform-es3-property-literals": "^6.22.0",
+    "css-loader": "^0.28.10",
+    "html-loader": "^0.5.5",
+    "html-webpack-plugin": "^3.0.4",
+    "postcss-loader": "^2.1.1",
+    "postcss-simple-vars": "^4.1.0",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
+    "style-loader": "^0.20.2",
+    "truffle-contract": "^3.0.4",
     "truffle-solidity-loader": "0.0.8",
-    "url-loader": "0.5.7",
-    "webpack": "1.14.0",
-    "webpack-dev-server": "1.16.2",
-    "webpack-manifest-plugin": "1.1.0",
-    "whatwg-fetch": "1.0.0"
+    "webpack": "^4.0.0",
+    "webpack-cli": "^2.0.8",
+    "webpack-dev-server": "^3.1.0"
   },
   "dependencies": {
-    "dotenv": "^2.0.0",
     "final-form": "^4.2.1",
     "material-ui": "^1.0.0-beta.35",
     "material-ui-icons": "^1.0.0-beta.35",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
-    "react-final-form": "^3.1.2"
-  },
-  "scripts": {
-    "start": "node scripts/start.js",
-    "build": "node scripts/build.js",
-    "test": "node scripts/test.js --env=jsdom"
-  },
-  "jest": {
-    "collectCoverageFrom": [
-      "src/**/*.{js,jsx}"
-    ],
-    "setupFiles": [
-      "<rootDir>/config/polyfills.js"
-    ],
-    "testMatch": [
-      "<rootDir>/src/**/__tests__/**/*.js?(x)",
-      "<rootDir>/src/**/?(*.)(spec|test).js?(x)"
-    ],
-    "testEnvironment": "node",
-    "testURL": "http://localhost",
-    "transform": {
-      "^.+\\.(js|jsx)$": "<rootDir>/node_modules/babel-jest",
-      "^.+\\.css$": "<rootDir>/config/jest/cssTransform.js",
-      "^(?!.*\\.(js|jsx|css|json)$)": "<rootDir>/config/jest/fileTransform.js"
-    },
-    "transformIgnorePatterns": [
-      "[/\\\\]node_modules[/\\\\].+\\.(js|jsx)$"
-    ],
-    "moduleNameMapper": {
-      "^react-native$": "react-native-web"
-    },
-    "moduleFileExtensions": [
-      "web.js",
-      "js",
-      "json",
-      "web.jsx",
-      "jsx"
-    ]
-  },
-  "babel": {
-    "presets": [
-      "react-app"
-    ]
-  },
-  "eslintConfig": {
-    "extends": "react-app"
+    "react-final-form": "^3.1.2",
+    "web3": "^1.0.0-beta.30"
   }
 }
